+++ packages/lib-react-components/test/setup.js
@@ -1,6 +1,5 @@
 /* eslint-env browser, mocha */
 /* eslint import/no-extraneous-dependencies: ["error", { "devDependencies": true  }] */
-import React from 'react'
 import chai from 'chai'
 import dirtyChai from 'dirty-chai'
 import sinonChai from 'sinon-chai'
@@ -10,7 +9,6 @@ import Adapter from '@wojtekmaj/enzyme-adapter-react-17'
 
 chai.use(dirtyChai)
 chai.use(sinonChai)
-global.React = React
 global.expect = chai.expect
 
 const jsdom = new JSDOM('<!doctype html><html><body></body></html>', { url: 'https://localhost'})