+++ packages/lib-react-components/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased 2023-09-22
+
+### Fixed
+- replace `require` with `import`.
+- remove the `React` global from tests.
+
 ## [1.6.2] 2023-09-19
 
 ### Fixed